Type missmatch for std::max (#74)
std::max is a template function that its parameters are the same type. The problem class is a template class, but if float type is used, there is call to std::max where float and double are compared since a double literal is used there. Fixed with static_cast as already done several lines below that bug.