Fix an off-by-one error in NelderMeadSolver
f_r should be compared to f_n and not f_{n+1} do decide if the
reflection point is accepted.

A side-effect of this bug was that there were only inside
contractions and no outside contractions as the condition for the
outside contraction was always false.

FYI: As this code is only executed if DIM >= 1 there shouldn't be
any accidental accesses to index -1.